Remove per-chapter PO files
Now the first round of the translation is completed and we now enter
the maintenance phase. ja.po is the master file and it makes
the situation complicated if we have two contents in the repo
(all-in-one ja.po and per-chapter PO files), so I remove per-chapter
PO files by this commit.

Note that per-chapter PO files can be regenerated by running the
following command in doc/src/docbkx/openstack-ops/locale:

  ./tools/chap-split.sh

Change-Id: Icfdbcd1d9249d15dec24d6ea80a337776977fe7d